refactor(webapp): archive requests page (#73)
* refactor: stomp client class

move it into a separate module

* refactor: separate archive requests page

* refactor: archival event page

* refactor: downloads in progress page

* fix: build errors

* style: misc fixes

* refactor: global header component

* refactor: archives root page
