+++ webapp/package.json
@@ -22,7 +22,7 @@
     "clsx": "^1.1.1",
     "form-data": "^4.0.0",
     "js-cookie": "^3.0.1",
-    "next": "^12.2.0",
+    "next": "12.2.0",
     "react": "^17.0.2",
     "react-context-mutex": "^2.0.0",
     "react-dom": "^17.0.2",
