+++ webapp/src/components/archives/_index.ts
@@ -1,0 +1,1 @@
+export { ArchiveRequestStatus as ArchiveStatus } from "./status";
