+++ webapp/src/components/header/_index.ts
@@ -1,0 +1,1 @@
+export { Header } from "./header";
