+++ webapp/src/lib/fetch/_index.ts
@@ -1,0 +1,1 @@
+export { WSClient, WSConfig } from "./websocket";
