+++ webapp/src/pages/account/archives/events.module.scss
@@ -1,0 +1,7 @@
+.heading {
+  text-align: center;
+  font-size: 2.25rem;
+  line-height: 2.5rem;
+  color: var(--colour0-primary);
+  margin: 0;
+}
