+++ webapp/src/pages/account/archives/requests.module.scss
@@ -1,0 +1,7 @@
+.heading {
+  text-align: center;
+  font-size: 1.25rem;
+  line-height: 1.75rem;
+  color: var(--colour0-primary);
+  margin-bottom: 1.25rem;
+}
